backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.2.1
Pygments==2.19.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
asyncpg==0.30.0
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Monitoring and metrics
prometheus-client==0.19.0
